+++ tests/dmrs_test.py
@@ -1,6 +1,7 @@
 
 import pytest
 
+from delphin.codecs import simplemrs
 from delphin import dmrs
 
 
@@ -137,3 +138,69 @@ class TestDMRS():
             10001: [('RSTR', 'qeq', scopemap[10002])],
             10002: []
         }
+
+
+def test_from_mrs_it_rains():
+    m = simplemrs.decode('''
+        [ TOP: h0 INDEX: e2 [e TENSE: pres]
+          RELS: < [ _rain_v_1<3:8> LBL: h1 ARG0: e2 ] >
+          HCONS: < h0 qeq h1 > ]''')
+    d = dmrs.from_mrs(m)
+    assert len(d.nodes) == 1
+    assert d.nodes[0].predicate == '_rain_v_1'
+    assert d.nodes[0].type == 'e'
+    assert d.nodes[0].properties == {'TENSE': 'pres'}
+
+
+def test_from_mrs_nearly_all_cats_were_chased_by_dogs():
+    m = simplemrs.decode('''
+        [ LTOP: h0
+          INDEX: e2 [ e SF: prop TENSE: past MOOD: indicative PROG: - PERF: - ]
+          RELS: < [ _nearly_x_deg<0:6> LBL: h4 ARG0: e5 [ e SF: prop TENSE: untensed MOOD: indicative PROG: - PERF: - ] ARG1: u6 ]
+                  [ _all_q<7:10> LBL: h4 ARG0: x3 [ x PERS: 3 NUM: pl IND: + ] RSTR: h7 BODY: h8 ]
+                  [ _cat_n_1<11:15> LBL: h9 ARG0: x3 ]
+                  [ _chase_v_1<21:27> LBL: h1 ARG0: e2 ARG1: x10 [ x PERS: 3 NUM: pl IND: + ] ARG2: x3 ]
+                  [ udef_q<31:36> LBL: h11 ARG0: x10 RSTR: h12 BODY: h13 ]
+                  [ _dog_n_1<31:36> LBL: h14 ARG0: x10 ] >
+          HCONS: < h0 qeq h1 h7 qeq h9 h12 qeq h14 >
+          ICONS: < e2 topic x3 > ]''')
+    d = dmrs.from_mrs(m)
+    assert len(d.nodes) == 6
+    n1 = d.nodes[0]
+    n2 = d.nodes[1]
+    assert n1.predicate == '_nearly_x_deg'
+    assert n2.predicate == '_all_q'
+    assert any((l.start, l.end, l.role, l.post) == (n1.id, n2.id, 'MOD', 'EQ')
+               for l in d.links)
+
+
+def test_from_mrs_issue_303():
+    # https://github.com/delph-in/pydelphin/issues/303
+    m = simplemrs.decode('''
+        [ TOP: h0 INDEX: e2 [e TENSE: pres]
+          RELS: < [ _rain_v_1<3:8> LBL: h1 ARG0: e2 ] >
+          HCONS: < > ]''')
+    with pytest.warns(dmrs.DMRSWarning):
+        d = dmrs.from_mrs(m)
+        assert d.top is None
+
+    m = simplemrs.decode('''
+        [ TOP: h0 INDEX: e2 [e TENSE: pres]
+          RELS: < [ _rain_v_1<3:8> LBL: h1 ARG0: e2 ] >
+          HCONS: < h0 qeq h3 > ]''')
+    with pytest.warns(dmrs.DMRSWarning):
+        d = dmrs.from_mrs(m)
+        assert d.top is None
+
+    m = simplemrs.decode('''
+        [ LTOP: h0
+          INDEX: e2 [ e SF: prop TENSE: past MOOD: indicative PROG: - PERF: - ]
+          RELS: < [ neg<7:10> LBL: h1 ARG0: e4 [ e SF: prop TENSE: untensed MOOD: indicative PROG: - PERF: - ] ARG1: h5 ]
+                  [ _rain_v_1<11:16> LBL: h6 ARG0: e2 ] >
+          HCONS: < h0 qeq h1 h5 qeq h7 >
+          ICONS: < > ]''')
+    with pytest.warns(dmrs.DMRSWarning):
+        d = dmrs.from_mrs(m)
+        n = d.nodes[0]
+        assert n.predicate == 'neg'
+        assert 'ARG1' not in d.scopal_arguments()[n.id]